Ex03- Advanced Math version 4

+++ src/com/mta/javacourse/Gal_NasdaqServlet.java
@@ -22,8 +22,8 @@ public class Gal_NasdaqServlet extends HttpServlet {
 		
 		String line1= new String ("<h1> Calculation 1: Area of circle with radius: "+radius+" * "+pi+" = "+result1+" </h1>");	
 		
-		int opposite;
-		int hypotenuse;
+		double opposite;
+		double hypotenuse;
 			
 		opposite=30;
 		hypotenuse=50;
@@ -31,8 +31,8 @@ public class Gal_NasdaqServlet extends HttpServlet {
 		
 		String line2= new String ("<h1> Calculation 2: Length of opposite where angle B is 30 degrees and Hypotenuse length is 50 cm is: "+opposite+" / "+hypotenuse+" = "+result2+"  cm </h1>");	
 		
-		int base;
-		int exp;
+		double base;
+		double exp;
 		
 		base=20;
 		exp=13;
